split runner.rs into smaller files
- this refactoring is in preparation for the next steps (commits on this
  branch)
- smaller files makes the scope of the objects easier to reason about,
  grouping them by ocptv topic
