+++ src/output/mod.rs
@@ -4,12 +4,17 @@
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+mod config;
 mod emitters;
 mod macros;
+mod measurement_series;
 mod models;
 mod objects;
-mod runner;
+mod run;
+mod state;
+mod step;
 
+pub use config::*;
 pub use emitters::*;
 pub use models::LogSeverity;
 pub use models::TestResult;
@@ -17,5 +22,6 @@ pub use models::TestStatus;
 pub use models::ValidatorType;
 pub use models::SPEC_VERSION;
 pub use objects::*;
-pub use runner::*;
+pub use run::*;
 pub use serde_json::Value;
+pub use step::*;
