--- src/output/runner.rs
@@ -1,1291 +1,0 @@
-// (c) Meta Platforms, Inc. and affiliates.
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
-
-//! OCPTV library runner
-//!
-//! This module contains the main entry point for the test runner. This is the
-//! main object the user will interact with.
-
-use std::env;
-use std::future::Future;
-use std::path::Path;
-use std::sync::atomic;
-use std::sync::Arc;
-
-use serde_json::Map;
-use serde_json::Value;
-use tokio::sync::Mutex;
-
-use crate::output::emitters;
-use crate::output::models;
-use crate::output::objects;
-
-/// The configuration repository for the TestRun.
-pub struct Config {
-    timezone: chrono_tz::Tz,
-    writer: emitters::WriterType,
-}
-
-impl Config {
-    /// Creates a new [`ConfigBuilder`]
-    ///
-    /// # Examples
-    /// ```rust
-    /// # use ocptv::output::*;
-    ///
-    /// let builder = Config::builder();
-    /// ```
-    pub fn builder() -> ConfigBuilder {
-        ConfigBuilder::new()
-    }
-}
-
-/// The builder for the [`Config`] object.
-pub struct ConfigBuilder {
-    timezone: Option<chrono_tz::Tz>,
-    writer: Option<emitters::WriterType>,
-}
-
-impl ConfigBuilder {
-    fn new() -> Self {
-        Self {
-            timezone: None,
-            writer: Some(emitters::WriterType::Stdout(emitters::StdoutWriter::new())),
-        }
-    }
-
-    pub fn timezone(mut self, timezone: chrono_tz::Tz) -> Self {
-        self.timezone = Some(timezone);
-        self
-    }
-
-    pub fn with_buffer_output(mut self, buffer: Arc<Mutex<Vec<String>>>) -> Self {
-        self.writer = Some(emitters::WriterType::Buffer(emitters::BufferWriter::new(
-            buffer,
-        )));
-        self
-    }
-
-    pub async fn with_file_output<P: AsRef<Path>>(
-        mut self,
-        path: P,
-    ) -> Result<Self, emitters::WriterError> {
-        self.writer = Some(emitters::WriterType::File(
-            emitters::FileWriter::new(path).await?,
-        ));
-        Ok(self)
-    }
-
-    pub fn build(self) -> Config {
-        Config {
-            timezone: self.timezone.unwrap_or(chrono_tz::UTC),
-            writer: self
-                .writer
-                .unwrap_or(emitters::WriterType::Stdout(emitters::StdoutWriter::new())),
-        }
-    }
-}
-
-/// The outcome of a TestRun.
-/// It's returned when the scope method of the [`TestRun`] object is used.
-pub struct TestRunOutcome {
-    /// Reports the execution status of the test
-    pub status: models::TestStatus,
-    /// Reports the result of the test
-    pub result: models::TestResult,
-}
-
-struct TestState {
-    emitter: emitters::JsonEmitter,
-}
-
-impl TestState {
-    fn new(emitter: emitters::JsonEmitter) -> TestState {
-        TestState { emitter }
-    }
-}
-
-/// The main diag test run.
-///
-/// This object describes a single run instance of the diag, and therefore drives the test session.
-pub struct TestRun {
-    name: String,
-    version: String,
-    parameters: Map<String, Value>,
-    dut: objects::DutInfo,
-    command_line: String,
-    metadata: Option<Map<String, Value>>,
-    state: Arc<Mutex<TestState>>,
-}
-
-impl TestRun {
-    /// Creates a new [`TestRunBuilder`] object.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use ocptv::output::*;
-    ///
-    /// let dut = DutInfo::builder("my_dut").build();
-    /// let builder = TestRun::builder("run_name", &dut, "1.0");
-    /// ```
-    pub fn builder(name: &str, dut: &objects::DutInfo, version: &str) -> TestRunBuilder {
-        TestRunBuilder::new(name, dut, version)
-    }
-
-    /// Creates a new [`TestRun`] object.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0");
-    /// ```
-    pub fn new(name: &str, dut_id: &str, version: &str) -> TestRun {
-        let dut = objects::DutInfo::new(dut_id);
-        TestRunBuilder::new(name, &dut, version).build()
-    }
-
-    /// Starts the test run.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#schemaversion
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#testrunstart
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0");
-    /// run.start().await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn start(self) -> Result<StartedTestRun, emitters::WriterError> {
-        let version = objects::SchemaVersion::new();
-        self.state
-            .lock()
-            .await
-            .emitter
-            .emit(&version.to_artifact())
-            .await?;
-
-        let mut builder = objects::TestRunStart::builder(
-            &self.name,
-            &self.version,
-            &self.command_line,
-            &self.parameters,
-            &self.dut,
-        );
-
-        if let Some(m) = &self.metadata {
-            for m in m {
-                builder = builder.add_metadata(m.0, m.1.clone())
-            }
-        }
-
-        let start = builder.build();
-        self.state
-            .lock()
-            .await
-            .emitter
-            .emit(&start.to_artifact())
-            .await?;
-
-        Ok(StartedTestRun { run: self })
-    }
-
-    // disabling this for the moment so we don't publish api that's unusable.
-    // see: https://github.com/rust-lang/rust/issues/70263
-    //
-    // /// Builds a scope in the [`TestRun`] object, taking care of starting and
-    // /// ending it. View [`TestRun::start`] and [`TestRun::end`] methods.
-    // /// After the scope is constructed, additional objects may be added to it.
-    // /// This is the preferred usage for the [`TestRun`], since it guarantees
-    // /// all the messages are emitted between the start and end messages, the order
-    // /// is respected and no messages is lost.
-    // ///
-    // /// # Examples
-    // ///
-    // /// ```rust
-    // /// # tokio_test::block_on(async {
-    // /// # use ocptv::output::*;
-    // ///
-    // /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0");
-    // /// run.scope(|r| async {
-    // ///     r.log(LogSeverity::Info, "First message").await?;
-    // ///     Ok(TestRunOutcome {
-    // ///         status: TestStatus::Complete,
-    // ///         result: TestResult::Pass,
-    // ///     })
-    // /// }).await?;
-    // ///
-    // /// # Ok::<(), WriterError>(())
-    // /// # });
-    // /// ```
-    // pub async fn scope<F, R>(self, func: F) -> Result<(), emitters::WriterError>
-    // where
-    //     R: Future<Output = Result<TestRunOutcome, emitters::WriterError>>,
-    //     for<'a> F: Fut2<'a, R>,
-    // {
-    //     let run = self.start().await?;
-    //     let outcome = func(&run).await?;
-    //     run.end(outcome.status, outcome.result).await?;
-
-    //     Ok(())
-    // }
-}
-
-/// Builder for the [`TestRun`] object.
-pub struct TestRunBuilder {
-    name: String,
-    dut: objects::DutInfo,
-    version: String,
-    parameters: Map<String, Value>,
-    command_line: String,
-    metadata: Option<Map<String, Value>>,
-    config: Option<Config>,
-}
-
-impl TestRunBuilder {
-    pub fn new(name: &str, dut: &objects::DutInfo, version: &str) -> Self {
-        Self {
-            name: name.to_string(),
-            dut: dut.clone(),
-            version: version.to_string(),
-            parameters: Map::new(),
-            command_line: env::args().collect::<Vec<_>>()[1..].join(" "),
-            metadata: None,
-            config: None,
-        }
-    }
-
-    /// Adds a user defined parameter to the future [`TestRun`] object.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use ocptv::output::*;
-    ///
-    /// let dut = DutInfo::builder("dut_id").build();
-    /// let run = TestRunBuilder::new("run_name", &dut, "1.0")
-    ///     .add_parameter("param1", "value1".into())
-    ///     .build();
-    /// ```
-    pub fn add_parameter(mut self, key: &str, value: Value) -> TestRunBuilder {
-        self.parameters.insert(key.to_string(), value.clone());
-        self
-    }
-
-    /// Adds the command line used to run the test session  to the future
-    /// [`TestRun`] object.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use ocptv::output::*;
-    ///
-    /// let dut = DutInfo::builder("dut_id").build();
-    /// let run = TestRunBuilder::new("run_name", &dut, "1.0")
-    ///     .command_line("my_diag --arg value")
-    ///     .build();
-    /// ```
-    pub fn command_line(mut self, cmd: &str) -> TestRunBuilder {
-        self.command_line = cmd.to_string();
-        self
-    }
-
-    /// Adds the configuration for the test session to the future [`TestRun`] object
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use ocptv::output::{Config, TestRunBuilder, DutInfo};
-    ///
-    /// let dut = DutInfo::builder("dut_id").build();
-    /// let run = TestRunBuilder::new("run_name", &dut, "1.0")
-    ///     .config(Config::builder().build())
-    ///     .build();
-    /// ```
-    pub fn config(mut self, value: Config) -> TestRunBuilder {
-        self.config = Some(value);
-        self
-    }
-
-    /// Adds user defined metadata to the future [`TestRun`] object
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use ocptv::output::*;
-    ///
-    /// let dut = DutInfo::builder("dut_id").build();
-    /// let run = TestRunBuilder::new("run_name", &dut, "1.0")
-    ///     .add_metadata("meta1", "value1".into())
-    ///     .build();
-    /// ```
-    pub fn add_metadata(mut self, key: &str, value: Value) -> TestRunBuilder {
-        self.metadata = match self.metadata {
-            Some(mut metadata) => {
-                metadata.insert(key.to_string(), value.clone());
-                Some(metadata)
-            }
-            None => {
-                let mut metadata = Map::new();
-                metadata.insert(key.to_string(), value.clone());
-                Some(metadata)
-            }
-        };
-        self
-    }
-
-    pub fn build(self) -> TestRun {
-        let config = self.config.unwrap_or(Config::builder().build());
-        let emitter = emitters::JsonEmitter::new(config.timezone, config.writer);
-        let state = TestState::new(emitter);
-        TestRun {
-            name: self.name,
-            dut: self.dut,
-            version: self.version,
-            parameters: self.parameters,
-            command_line: self.command_line,
-            metadata: self.metadata,
-            state: Arc::new(Mutex::new(state)),
-        }
-    }
-}
-
-/// A test run that was started.
-///
-/// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#testrunstart
-pub struct StartedTestRun {
-    run: TestRun,
-}
-
-impl StartedTestRun {
-    /// Ends the test run.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#testrunend
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// run.end(TestStatus::Complete, TestResult::Pass).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn end(
-        &self,
-        status: models::TestStatus,
-        result: models::TestResult,
-    ) -> Result<(), emitters::WriterError> {
-        let end = objects::TestRunEnd::builder()
-            .status(status)
-            .result(result)
-            .build();
-
-        let emitter = &self.run.state.lock().await.emitter;
-
-        emitter.emit(&end.to_artifact()).await?;
-        Ok(())
-    }
-
-    /// Emits a Log message.
-    /// This method accepts a [`models::LogSeverity`] to define the severity
-    /// and a [`std::string::String`] for the message.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#log
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// run.log(
-    ///     LogSeverity::Info,
-    ///     "This is a log message with INFO severity",
-    /// ).await?;
-    /// run.end(TestStatus::Complete, TestResult::Pass).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn log(
-        &self,
-        severity: models::LogSeverity,
-        msg: &str,
-    ) -> Result<(), emitters::WriterError> {
-        let log = objects::Log::builder(msg).severity(severity).build();
-
-        let emitter = &self.run.state.lock().await.emitter;
-
-        emitter
-            .emit(&log.to_artifact(objects::ArtifactContext::TestRun))
-            .await?;
-        Ok(())
-    }
-
-    /// Emits a Log message.
-    /// This method accepts a [`objects::Log`] object.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#log
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// run.log_with_details(
-    ///     &Log::builder("This is a log message with INFO severity")
-    ///         .severity(LogSeverity::Info)
-    ///         .source("file", 1)
-    ///         .build(),
-    /// ).await?;
-    /// run.end(TestStatus::Complete, TestResult::Pass).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn log_with_details(&self, log: &objects::Log) -> Result<(), emitters::WriterError> {
-        let emitter = &self.run.state.lock().await.emitter;
-
-        emitter
-            .emit(&log.to_artifact(objects::ArtifactContext::TestRun))
-            .await?;
-        Ok(())
-    }
-
-    /// Emits a Error message.
-    /// This method accepts a [`std::string::String`] to define the symptom.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#error
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// run.error("symptom").await?;
-    /// run.end(TestStatus::Complete, TestResult::Pass).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn error(&self, symptom: &str) -> Result<(), emitters::WriterError> {
-        let error = objects::Error::builder(symptom).build();
-        let emitter = &self.run.state.lock().await.emitter;
-
-        emitter
-            .emit(&error.to_artifact(objects::ArtifactContext::TestRun))
-            .await?;
-        Ok(())
-    }
-
-    /// Emits a Error message.
-    /// This method accepts a [`std::string::String`] to define the symptom and
-    /// another [`std::string::String`] as error message.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#error
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// run.error_with_msg("symptom", "error messasge").await?;
-    /// run.end(TestStatus::Complete, TestResult::Pass).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn error_with_msg(
-        &self,
-        symptom: &str,
-        msg: &str,
-    ) -> Result<(), emitters::WriterError> {
-        let error = objects::Error::builder(symptom).message(msg).build();
-        let emitter = &self.run.state.lock().await.emitter;
-
-        emitter
-            .emit(&error.to_artifact(objects::ArtifactContext::TestRun))
-            .await?;
-        Ok(())
-    }
-
-    /// Emits a Error message.
-    /// This method acceps a [`objects::Error`] object.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#error
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// run.error_with_details(
-    ///     &Error::builder("symptom")
-    ///         .message("Error message")
-    ///         .source("file", 1)
-    ///         .add_software_info(&SoftwareInfo::builder("id", "name").build())
-    ///         .build(),
-    /// ).await?;
-    /// run.end(TestStatus::Complete, TestResult::Pass).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn error_with_details(
-        &self,
-        error: &objects::Error,
-    ) -> Result<(), emitters::WriterError> {
-        let emitter = &self.run.state.lock().await.emitter;
-
-        emitter
-            .emit(&error.to_artifact(objects::ArtifactContext::TestRun))
-            .await?;
-        Ok(())
-    }
-
-    pub fn step(&self, name: &str) -> TestStep {
-        TestStep::new(name, self.run.state.clone())
-    }
-}
-
-/// A single test step in the scope of a [`TestRun`].
-///
-/// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#test-step-artifacts
-pub struct TestStep {
-    name: String,
-    state: Arc<Mutex<TestState>>,
-}
-
-impl TestStep {
-    fn new(name: &str, state: Arc<Mutex<TestState>>) -> TestStep {
-        TestStep {
-            name: name.to_string(),
-            state,
-        }
-    }
-
-    /// Starts the test step.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#teststepstart
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// let step = run.step("step_name").start().await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn start(self) -> Result<StartedTestStep, emitters::WriterError> {
-        let start = objects::TestStepStart::new(&self.name);
-        self.state
-            .lock()
-            .await
-            .emitter
-            .emit(&start.to_artifact())
-            .await?;
-
-        Ok(StartedTestStep {
-            step: self,
-            measurement_id_no: Arc::new(atomic::AtomicU64::new(0)),
-        })
-    }
-
-    // /// Builds a scope in the [`TestStep`] object, taking care of starting and
-    // /// ending it. View [`TestStep::start`] and [`TestStep::end`] methods.
-    // /// After the scope is constructed, additional objects may be added to it.
-    // /// This is the preferred usage for the [`TestStep`], since it guarantees
-    // /// all the messages are emitted between the start and end messages, the order
-    // /// is respected and no messages is lost.
-    // ///
-    // /// # Examples
-    // ///
-    // /// ```rust
-    // /// # tokio_test::block_on(async {
-    // /// # use ocptv::output::*;
-    // ///
-    // /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    // ///
-    // /// let step = run.step("first step")?;
-    // /// step.scope(|s| async {
-    // ///     s.log(
-    // ///         LogSeverity::Info,
-    // ///         "This is a log message with INFO severity",
-    // ///     ).await?;
-    // ///     Ok(TestStatus::Complete)
-    // /// }).await?;
-    // ///
-    // /// # Ok::<(), WriterError>(())
-    // /// # });
-    // /// ```
-    // pub async fn scope<'a, F, R>(&'a self, func: F) -> Result<(), emitters::WriterError>
-    // where
-    //     R: Future<Output = Result<models::TestStatus, emitters::WriterError>>,
-    //     F: std::ops::FnOnce(&'a TestStep) -> R,
-    // {
-    //     self.start().await?;
-    //     let status = func(self).await?;
-    //     self.end(status).await?;
-    //     Ok(())
-    // }
-}
-
-pub struct StartedTestStep {
-    step: TestStep,
-    measurement_id_no: Arc<atomic::AtomicU64>,
-}
-
-impl StartedTestStep {
-    /// Ends the test step.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#teststepend
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    ///
-    /// let step = run.step("step_name").start().await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn end(&self, status: models::TestStatus) -> Result<(), emitters::WriterError> {
-        let end = objects::TestStepEnd::new(status);
-        self.step
-            .state
-            .lock()
-            .await
-            .emitter
-            .emit(&end.to_artifact())
-            .await?;
-        Ok(())
-    }
-
-    /// Eemits Log message.
-    /// This method accepts a [`models::LogSeverity`] to define the severity
-    /// and a [`std::string::String`] for the message.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#log
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    ///
-    /// let step = run.step("step_name").start().await?;
-    /// step.log(
-    ///     LogSeverity::Info,
-    ///     "This is a log message with INFO severity",
-    /// ).await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    /// ## Using macros
-    ///  
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// use ocptv::ocptv_log_info;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    ///
-    /// let step = run.step("step_name").start().await?;
-    /// ocptv_log_info!(step, "This is a log message with INFO severity").await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn log(
-        &self,
-        severity: models::LogSeverity,
-        msg: &str,
-    ) -> Result<(), emitters::WriterError> {
-        let log = objects::Log::builder(msg).severity(severity).build();
-        self.step
-            .state
-            .lock()
-            .await
-            .emitter
-            .emit(&log.to_artifact(objects::ArtifactContext::TestStep))
-            .await?;
-        Ok(())
-    }
-
-    /// Emits Log message.
-    /// This method accepts a [`objects::Log`] object.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#log
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    ///
-    /// let step = run.step("step_name").start().await?;
-    /// step.log_with_details(
-    ///     &Log::builder("This is a log message with INFO severity")
-    ///         .severity(LogSeverity::Info)
-    ///         .source("file", 1)
-    ///         .build(),
-    /// ).await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn log_with_details(&self, log: &objects::Log) -> Result<(), emitters::WriterError> {
-        self.step
-            .state
-            .lock()
-            .await
-            .emitter
-            .emit(&log.to_artifact(objects::ArtifactContext::TestStep))
-            .await?;
-        Ok(())
-    }
-
-    /// Emits an Error symptom.
-    /// This method accepts a [`std::string::String`] to define the symptom.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#error
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    ///
-    /// let step = run.step("step_name").start().await?;
-    /// step.error("symptom").await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    ///
-    /// ## Using macros
-    ///  
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// use ocptv::ocptv_error;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    ///
-    /// let step = run.step("step_name").start().await?;
-    /// ocptv_error!(step, "symptom").await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn error(&self, symptom: &str) -> Result<(), emitters::WriterError> {
-        let error = objects::Error::builder(symptom).build();
-        self.step
-            .state
-            .lock()
-            .await
-            .emitter
-            .emit(&error.to_artifact(objects::ArtifactContext::TestStep))
-            .await?;
-        Ok(())
-    }
-
-    /// Emits an Error message.
-    /// This method accepts a [`std::string::String`] to define the symptom and
-    /// another [`std::string::String`] as error message.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#error
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    ///
-    /// let step = run.step("step_name").start().await?;
-    /// step.error_with_msg("symptom", "error message").await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    ///
-    /// ## Using macros
-    ///  
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// use ocptv::ocptv_error;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    ///
-    /// let step = run.step("step_name").start().await?;
-    /// ocptv_error!(step, "symptom", "error message").await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn error_with_msg(
-        &self,
-        symptom: &str,
-        msg: &str,
-    ) -> Result<(), emitters::WriterError> {
-        let error = objects::Error::builder(symptom).message(msg).build();
-        self.step
-            .state
-            .lock()
-            .await
-            .emitter
-            .emit(&error.to_artifact(objects::ArtifactContext::TestStep))
-            .await?;
-        Ok(())
-    }
-
-    /// Emits a Error message.
-    /// This method accepts a [`objects::Error`] object.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#error
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    ///
-    /// let step = run.step("step_name").start().await?;
-    /// step.error_with_details(
-    ///     &Error::builder("symptom")
-    ///         .message("Error message")
-    ///         .source("file", 1)
-    ///         .add_software_info(&SoftwareInfo::builder("id", "name").build())
-    ///         .build(),
-    /// ).await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn error_with_details(
-        &self,
-        error: &objects::Error,
-    ) -> Result<(), emitters::WriterError> {
-        self.step
-            .state
-            .lock()
-            .await
-            .emitter
-            .emit(&error.to_artifact(objects::ArtifactContext::TestStep))
-            .await?;
-        Ok(())
-    }
-
-    /// Emits a Measurement message.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#measurement
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    ///
-    /// let step = run.step("step_name").start().await?;
-    /// step.add_measurement("name", 50.into()).await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn add_measurement(
-        &self,
-        name: &str,
-        value: Value,
-    ) -> Result<(), emitters::WriterError> {
-        let measurement = objects::Measurement::new(name, value);
-        self.step
-            .state
-            .lock()
-            .await
-            .emitter
-            .emit(&measurement.to_artifact())
-            .await?;
-        Ok(())
-    }
-
-    /// Emits a Measurement message.
-    /// This method accepts a [`objects::Error`] object.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#measurement
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let hwinfo = HardwareInfo::builder("id", "fan").build();
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// let step = run.step("step_name").start().await?;
-    ///
-    /// let measurement = Measurement::builder("name", 5000.into())
-    ///     .hardware_info(&hwinfo)
-    ///     .add_validator(&Validator::builder(ValidatorType::Equal, 30.into()).build())
-    ///     .add_metadata("key", "value".into())
-    ///     .subcomponent(&Subcomponent::builder("name").build())
-    ///     .build();
-    /// step.add_measurement_with_details(&measurement).await?;
-    /// step.end(TestStatus::Complete).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn add_measurement_with_details(
-        &self,
-        measurement: &objects::Measurement,
-    ) -> Result<(), emitters::WriterError> {
-        self.step
-            .state
-            .lock()
-            .await
-            .emitter
-            .emit(&measurement.to_artifact())
-            .await?;
-        Ok(())
-    }
-
-    /// Starts a Measurement Series (a time-series list of measurements).
-    /// This method accepts a [`std::string::String`] as series ID and
-    /// a [`std::string::String`] as series name.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#measurementseriesstart
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// let step = run.step("step_name").start().await?;
-    /// let series = step.measurement_series("name");
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub fn measurement_series(&self, name: &str) -> MeasurementSeries {
-        self.measurement_id_no
-            .fetch_add(1, atomic::Ordering::SeqCst);
-        let series_id: String = format!(
-            "series_{}",
-            self.measurement_id_no.load(atomic::Ordering::SeqCst)
-        );
-
-        MeasurementSeries::new(&series_id, name, self.step.state.clone())
-    }
-
-    /// Starts a Measurement Series (a time-series list of measurements).
-    /// This method accepts a [`objects::MeasurementSeriesStart`] object.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#measurementseriesstart
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// let step = run.step("step_name").start().await?;
-    /// let series =
-    ///     step.measurement_series_with_details(MeasurementSeriesStart::new("name", "series_id"));
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub fn measurement_series_with_details(
-        &self,
-        start: objects::MeasurementSeriesStart,
-    ) -> MeasurementSeries {
-        MeasurementSeries::new_with_details(start, self.step.state.clone())
-    }
-}
-
-/// The measurement series.
-/// A Measurement Series is a time-series list of measurements.
-///
-/// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#measurementseriesstart
-pub struct MeasurementSeries {
-    state: Arc<Mutex<TestState>>,
-    seq_no: Arc<Mutex<atomic::AtomicU64>>,
-    start: objects::MeasurementSeriesStart,
-}
-
-impl MeasurementSeries {
-    fn new(series_id: &str, name: &str, state: Arc<Mutex<TestState>>) -> Self {
-        Self {
-            state,
-            seq_no: Arc::new(Mutex::new(atomic::AtomicU64::new(0))),
-            start: objects::MeasurementSeriesStart::new(name, series_id),
-        }
-    }
-
-    fn new_with_details(
-        start: objects::MeasurementSeriesStart,
-        state: Arc<Mutex<TestState>>,
-    ) -> Self {
-        Self {
-            state,
-            seq_no: Arc::new(Mutex::new(atomic::AtomicU64::new(0))),
-            start,
-        }
-    }
-
-    async fn current_sequence_no(&self) -> u64 {
-        self.seq_no.lock().await.load(atomic::Ordering::SeqCst)
-    }
-
-    async fn increment_sequence_no(&self) {
-        self.seq_no
-            .lock()
-            .await
-            .fetch_add(1, atomic::Ordering::SeqCst);
-    }
-
-    /// Starts the measurement series.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#measurementseriesstart
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// let step = run.step("step_name").start().await?;
-    ///
-    /// let series = step.measurement_series("name");
-    /// series.start().await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn start(&self) -> Result<(), emitters::WriterError> {
-        self.state
-            .lock()
-            .await
-            .emitter
-            .emit(&self.start.to_artifact())
-            .await?;
-        Ok(())
-    }
-
-    /// Ends the measurement series.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#measurementseriesend
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// let step = run.step("step_name").start().await?;
-    ///
-    /// let series = step.measurement_series("name");
-    /// series.start().await?;
-    /// series.end().await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn end(&self) -> Result<(), emitters::WriterError> {
-        let end = objects::MeasurementSeriesEnd::new(
-            self.start.get_series_id(),
-            self.current_sequence_no().await,
-        );
-        self.state
-            .lock()
-            .await
-            .emitter
-            .emit(&end.to_artifact())
-            .await?;
-        Ok(())
-    }
-
-    /// Adds a measurement element to the measurement series.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#measurementserieselement
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// let step = run.step("step_name").start().await?;
-    ///
-    /// let series = step.measurement_series("name");
-    /// series.start().await?;
-    /// series.add_measurement(60.into()).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn add_measurement(&self, value: Value) -> Result<(), emitters::WriterError> {
-        let element = objects::MeasurementSeriesElement::new(
-            self.current_sequence_no().await,
-            value,
-            &self.start,
-            None,
-        );
-        self.increment_sequence_no().await;
-        self.state
-            .lock()
-            .await
-            .emitter
-            .emit(&element.to_artifact())
-            .await?;
-        Ok(())
-    }
-
-    /// Adds a measurement element to the measurement series.
-    /// This method accepts additional metadata to add to the element.
-    ///
-    /// ref: https://github.com/opencomputeproject/ocp-diag-core/tree/main/json_spec#measurementserieselement
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// let step = run.step("step_name").start().await?;
-    ///
-    /// let series = step.measurement_series("name");
-    /// series.start().await?;
-    /// series.add_measurement_with_metadata(60.into(), vec![("key", "value".into())]).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn add_measurement_with_metadata(
-        &self,
-        value: Value,
-        metadata: Vec<(&str, Value)>,
-    ) -> Result<(), emitters::WriterError> {
-        let element = objects::MeasurementSeriesElement::new(
-            self.current_sequence_no().await,
-            value,
-            &self.start,
-            Some(Map::from_iter(
-                metadata.iter().map(|(k, v)| (k.to_string(), v.clone())),
-            )),
-        );
-        self.increment_sequence_no().await;
-        self.state
-            .lock()
-            .await
-            .emitter
-            .emit(&element.to_artifact())
-            .await?;
-        Ok(())
-    }
-
-    /// Builds a scope in the [`MeasurementSeries`] object, taking care of starting and
-    /// ending it. View [`MeasurementSeries::start`] and [`MeasurementSeries::end`] methods.
-    /// After the scope is constructed, additional objects may be added to it.
-    /// This is the preferred usage for the [`MeasurementSeries`], since it guarantees
-    /// all the messages are emitted between the start and end messages, the order
-    /// is respected and no messages is lost.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # tokio_test::block_on(async {
-    /// # use ocptv::output::*;
-    ///
-    /// let run = TestRun::new("diagnostic_name", "my_dut", "1.0").start().await?;
-    /// let step = run.step("step_name").start().await?;
-    ///
-    /// let series = step.measurement_series("name");
-    /// series.start().await?;
-    /// series.scope(|s| async {
-    ///     s.add_measurement(60.into()).await?;
-    ///     s.add_measurement(70.into()).await?;
-    ///     s.add_measurement(80.into()).await?;
-    ///     Ok(())
-    /// }).await?;
-    ///
-    /// # Ok::<(), WriterError>(())
-    /// # });
-    /// ```
-    pub async fn scope<'a, F, R>(&'a self, func: F) -> Result<(), emitters::WriterError>
-    where
-        R: Future<Output = Result<(), emitters::WriterError>>,
-        F: std::ops::FnOnce(&'a MeasurementSeries) -> R,
-    {
-        self.start().await?;
-        func(self).await?;
-        self.end().await?;
-        Ok(())
-    }
-}
